NTR: change tests
