+++ tests/Unit/Fake/FakeRouter.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mollie\Shopware\Unit\Fake;
+
+use Symfony\Component\Routing\RequestContext;
+use Symfony\Component\Routing\RouteCollection;
+use Symfony\Component\Routing\RouterInterface;
+
+final class FakeRouter implements RouterInterface
+{
+    private RequestContext $context;
+
+    public function __construct(private string $generatedUrl = '')
+    {
+        $this->context = new RequestContext();
+    }
+
+    public function generate(string $name, array $parameters = [], int $referenceType = self::ABSOLUTE_PATH): string
+    {
+        return $this->generatedUrl;
+    }
+
+    public function setContext(RequestContext $context): void
+    {
+        $this->context = $context;
+    }
+
+    public function getContext(): RequestContext
+    {
+        return $this->context;
+    }
+
+    public function getRouteCollection(): RouteCollection
+    {
+        return new RouteCollection();
+    }
+
+    public function match(string $pathinfo): array
+    {
+        return [];
+    }
+}
